requirements.txt
streamlit==1.32.2
openai
PyMuPDF
pydantic
Pillow
fpdf2
